refactor(application): move files to correct namespace
